TA-191 tests fixes

+++ core/src/test/groovy/com/arturjarosz/task/project/application/impl/CostApplicationServiceImplTest.groovy
@@ -219,7 +219,7 @@ class CostApplicationServiceImplTest extends Specification {
             1 * this.projectRepository.save({
                 Project project ->
                     Cost cost = project.getCosts().iterator().next();
-                    cost.getValue().doubleValue() == NEW_VALUE;
+                    cost.getValue() == NEW_VALUE;
                     cost.getName() == NEW_NAME;
                     cost.getNote() == NEW_NOTE;
                     cost.getDate() == NEW_DATE;
